Updated main.py 20/08/2024
New changes:
- Added a new sub option to the existing sub option "more options". It is a sub option to show the visualisation options
Issues:
Sometimes the new sub options wouldn't show up when selecting it as an option in the UI. Fixed it by making sure the variables that control the shown options are properly accounted for (True/False values).
Why this change?
To make the UI more organised.
Next changes:
Might add a new function to visualise the average prices for each month, if time is available to do so. Otherwise, focus on the testing and evaluation and peer review.